+++ src/mdx-pages/dwango-event.mdx
@@ -1,0 +1,7 @@
+---
+path: "/blog/dwango-event"
+date: "2019-11-06"
+title: "ドワンゴ見学エントリ"
+status: "private"
+tags: ["dev"]
+---
