+++ src/mdx-pages/my-zsh.mdx
@@ -3,6 +3,7 @@ path: "/blog/my-zshrc"
 date: "2019-10-26"
 title: "わがzsh環境"
 status: "draft"
+tags: ["dev"]
 ---
 
 # zshすき
