+++ src/mdx-pages/post-1.mdx
@@ -2,6 +2,7 @@
 path: "/blog/hello-world"
 date: "2019-10-21"
 title: "このサイトを書いてるあほについて"
+tags: ["もみにす", "kkrn.me"]
 ---
 
 # はじめに
