+++ src/templates/blogTemplate.jsx
@@ -1,10 +1,9 @@
 import React from "react"
 import { Link } from "gatsby"
 import { MDXRenderer } from "gatsby-plugin-mdx"
-import { SiteHeader } from "../components/HeaderFooter"
-import { BlogStyle } from "../components/style"
+import ArticleHead from "../components/blog/ArticleHead"
+import BlogWrapper from "../components/wrappers/BlogWrapper"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
-import Twemoji from "react-twemoji"
 import {
   faChevronLeft,
   faChevronRight,
@@ -15,48 +14,17 @@ export default function PageTemplate({ pageContext }) {
 
   const node = pageContext.node
   return (
-    <div>
-      <BlogStyle />
-      <SiteHeader></SiteHeader>
-      <Twemoji noWrapper>
-        <main>
-          {PrevNextLink(pageContext)}
-          <article css={{}}>
-            <h1
-              css={{
-                fontSize: `200%`,
-                display: `flex`,
-                justifyContent: `space-between`,
-                width: `100%`,
-                flexWrap: `wrap`,
-                "@media (width <= 800px)": {
-                  flexWrap: `wrap`,
-                },
-              }}
-            >
-              <span>{node.frontmatter.title} </span>
-              <span
-                css={{
-                  alignSelf: `flex-end`,
-                  fontSize: `75%`,
-                  fontWeight: `400`,
-                  flexGrow: 1,
-                  textAlign: `right`,
-                  "@media (width <= 800px)": {},
-                }}
-              >
-                Date:{node.frontmatter.date}
-              </span>
-            </h1>
-            <MDXRenderer>{node.body}</MDXRenderer>
-            {node.frontmatter.status === "draft" ? (
-              <p>(この記事は未完成、まだ更新中なんだ。すまない)</p>
-            ) : null}
-          </article>
-          {PrevNextLink(pageContext)}
-        </main>
-      </Twemoji>
-    </div>
+    <BlogWrapper>
+      {PrevNextLink(pageContext)}
+      <article css={{}}>
+        <ArticleHead node={node} />
+        <MDXRenderer>{node.body}</MDXRenderer>
+        {node.frontmatter.status === "draft" ? (
+          <p>(この記事は未完成、まだ更新中なんだ。すまない)</p>
+        ) : null}
+      </article>
+      {PrevNextLink(pageContext)}
+    </BlogWrapper>
   )
 }
